requirements.txt
# MCP DevOps Tools - Dependencies

# MCP SDK
mcp>=1.0.0

# HTTP requests
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
anyio>=4.0.0

# For Doris (MySQL protocol) - optional
# pymysql>=1.1.0
